--- a/cmd/otel-allocator/internal/allocation/allocator_test.go
+++ b/cmd/otel-allocator/internal/allocation/allocator_test.go
@@ -174,10 +174,7 @@ func TestAllocationCollision(t *testing.T) {
 		firstTarget := target.NewItem("sample-name", "0.0.0.0:8000", firstLabels, "")
 		secondTarget := target.NewItem("sample-name", "0.0.0.0:8000", secondLabels, "")
 
-		targetList := map[string]*target.Item{
-			firstTarget.Hash():  firstTarget,
-			secondTarget.Hash(): secondTarget,
-		}
+		targetList := []*target.Item{firstTarget, secondTarget}
 
 		// test that targets and collectors are added properly
 		allocator.SetTargets(targetList)