--- a/cmd/otel-allocator/internal/allocation/testutils.go
+++ b/cmd/otel-allocator/internal/allocation/testutils.go
@@ -24,8 +24,8 @@ func colIndex(index, numCols int) int {
 	return index % numCols
 }
 
-func MakeNNewTargets(n int, numCollectors int, startingIndex int) map[string]*target.Item {
-	toReturn := map[string]*target.Item{}
+func MakeNNewTargets(n int, numCollectors int, startingIndex int) []*target.Item {
+	toReturn := []*target.Item{}
 	for i := startingIndex; i < n+startingIndex; i++ {
 		collector := fmt.Sprintf("collector-%d", colIndex(i, numCollectors))
 		label := labels.Labels{
@@ -33,7 +33,7 @@ func MakeNNewTargets(n int, numCollectors int, startingIndex int) map[string]*ta
 			{Name: "total", Value: strconv.Itoa(n + startingIndex)},
 		}
 		newTarget := target.NewItem(fmt.Sprintf("test-job-%d", i), fmt.Sprintf("test-url-%d", i), label, collector)
-		toReturn[newTarget.Hash()] = newTarget
+		toReturn = append(toReturn, newTarget)
 	}
 	return toReturn
 }
@@ -51,8 +51,8 @@ func MakeNCollectors(n int, startingIndex int) map[string]*Collector {
 	return toReturn
 }
 
-func MakeNNewTargetsWithEmptyCollectors(n int, startingIndex int) map[string]*target.Item {
-	toReturn := map[string]*target.Item{}
+func MakeNNewTargetsWithEmptyCollectors(n int, startingIndex int) []*target.Item {
+	toReturn := []*target.Item{}
 	for i := startingIndex; i < n+startingIndex; i++ {
 		label := labels.Labels{
 			{Name: "i", Value: strconv.Itoa(i)},
@@ -60,7 +60,7 @@ func MakeNNewTargetsWithEmptyCollectors(n int, startingIndex int) map[string]*ta
 			{Name: "__meta_kubernetes_pod_node_name", Value: "node-0"},
 		}
 		newTarget := target.NewItem(fmt.Sprintf("test-job-%d", i), fmt.Sprintf("test-url-%d", i), label, "")
-		toReturn[newTarget.Hash()] = newTarget
+		toReturn = append(toReturn, newTarget)
 	}
 	return toReturn
 }