--- a/cmd/otel-allocator/internal/target/discovery.go
+++ b/cmd/otel-allocator/internal/target/discovery.go
@@ -54,8 +54,7 @@ type Discoverer struct {
 	scrapeConfigsUpdater   scrapeConfigsUpdater
 	targetSets             map[string][]*targetgroup.Group
 	triggerReload          chan struct{}
-	processTargetsCallBack func(targets map[string]*Item)
-	mtxTargets             sync.Mutex
+	processTargetsCallBack func(targets []*Item)
 }
 
 type discoveryHook interface {
@@ -66,7 +65,7 @@ type scrapeConfigsUpdater interface {
 	UpdateScrapeConfigResponse(map[string]*promconfig.ScrapeConfig) error
 }
 
-func NewDiscoverer(log logr.Logger, manager *discovery.Manager, hook discoveryHook, scrapeConfigsUpdater scrapeConfigsUpdater, setTargets func(targets map[string]*Item)) *Discoverer {
+func NewDiscoverer(log logr.Logger, manager *discovery.Manager, hook discoveryHook, scrapeConfigsUpdater scrapeConfigsUpdater, setTargets func(targets []*Item)) *Discoverer {
 	return &Discoverer{
 		log:                    log,
 		manager:                manager,
@@ -163,22 +162,29 @@ func (m *Discoverer) reloader() {
 func (m *Discoverer) Reload() {
 	m.mtxScrape.Lock()
 	var wg sync.WaitGroup
-	targets := map[string]*Item{}
 	timer := prometheus.NewTimer(processTargetsDuration)
 	defer timer.ObserveDuration()
 
+	// count targets and preallocate
+	targetCount := 0
+	for _, groups := range m.targetSets {
+		for _, group := range groups {
+			targetCount += len(group.Targets)
+		}
+	}
+	targets := make([]*Item, targetCount)
+
+	targetsAssigned := 0
 	for jobName, groups := range m.targetSets {
 		wg.Add(1)
 		// Run the sync in parallel as these take a while and at high load can't catch up.
-		go func(jobName string, groups []*targetgroup.Group) {
-			processedTargets := m.processTargetGroups(jobName, groups)
-			m.mtxTargets.Lock()
-			for k, v := range processedTargets {
-				targets[k] = v
-			}
-			m.mtxTargets.Unlock()
+		go func(jobName string, groups []*targetgroup.Group, intoTargets []*Item) {
+			m.processTargetGroups(jobName, groups, intoTargets)
 			wg.Done()
-		}(jobName, groups)
+		}(jobName, groups, targets[targetsAssigned:])
+		for _, group := range groups {
+			targetsAssigned += len(group.Targets)
+		}
 	}
 	m.mtxScrape.Unlock()
 	wg.Wait()
@@ -186,12 +192,13 @@ func (m *Discoverer) Reload() {
 }
 
 // processTargetGroups processes the target groups and returns a map of targets.
-func (m *Discoverer) processTargetGroups(jobName string, groups []*targetgroup.Group) map[string]*Item {
+func (m *Discoverer) processTargetGroups(jobName string, groups []*targetgroup.Group, intoTargets []*Item) {
 	builder := labels.NewBuilder(labels.Labels{})
 	timer := prometheus.NewTimer(processTargetGroupsDuration.WithLabelValues(jobName))
-	targets := map[string]*Item{}
+
 	defer timer.ObserveDuration()
 	var count float64 = 0
+	index := 0
 	for _, tg := range groups {
 		builder.Reset(labels.EmptyLabels())
 		for ln, lv := range tg.Labels {
@@ -205,11 +212,11 @@ func (m *Discoverer) processTargetGroups(jobName string, groups []*targetgroup.G
 				builder.Set(string(ln), string(lv))
 			}
 			item := NewItem(jobName, string(t[model.AddressLabel]), builder.Labels(), "")
-			targets[item.Hash()] = item
+			intoTargets[index] = item
+			index++
 		}
 	}
 	targetsDiscovered.WithLabelValues(jobName).Set(count)
-	return targets
 }
 
 // Run receives and saves target set updates and triggers the scraping loops reloading.