**Description:**

Instead of hashing scrape targets when they're received from service discovery, wait until the allocator actually needs the hash. This has the following benefits

* We can avoid calculating hashes for targets filtered out by relabeling.
* We hold targets in a slice, which is more performant.
* It becomes easier to solve issues which rely on accessing post-relabelling target state, like #3617.

**Link to tracking Issue(s):** <Issue number if applicable>

Makes it much easier to resolve #3617.

**Testing:**

Existing unit and e2e tests suffice.