@jaronoff97 2025-03-06
Looks good, I like this approach. I think it's fine to memoize the hash allocation as well, lgtm.