@swiatekm 2025-03-18
Ok, I think this is fine the way it is. We lose some performance when not using relabeling, but gain when we do. I'm not really sure what the reason for the regression is, but I don't want to hold this PR up with it, and we can improve performance further in various ways.

```
goos: linux
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator
cpu: AMD Ryzen 9 7950X3D 16-Core Processor          
                                                             │ bench_main.txt │          bench_branch.txt           │
                                                             │     sec/op     │   sec/op     vs base                │
ProcessTargets/least-weighted/1000-32                            1.037m ±  2%   1.060m ± 2%   +2.28% (p=0.019 n=10)
ProcessTargets/consistent-hashing/1000-32                        1.041m ±  2%   1.079m ± 3%   +3.61% (p=0.001 n=10)
ProcessTargets/per-node/1000-32                                  1.051m ±  2%   1.087m ± 2%   +3.35% (p=0.000 n=10)
ProcessTargets/least-weighted/10000-32                           7.052m ±  2%   7.563m ± 3%   +7.25% (p=0.000 n=10)
ProcessTargets/consistent-hashing/10000-32                       7.015m ±  3%   7.595m ± 6%   +8.27% (p=0.000 n=10)
ProcessTargets/per-node/10000-32                                 7.061m ±  3%   7.532m ± 4%   +6.68% (p=0.000 n=10)
ProcessTargets/least-weighted/100000-32                          67.77m ±  4%   84.76m ± 3%  +25.06% (p=0.000 n=10)
ProcessTargets/consistent-hashing/100000-32                      67.55m ±  6%   87.66m ± 6%  +29.77% (p=0.000 n=10)
ProcessTargets/per-node/100000-32                                68.54m ±  4%   81.04m ± 7%  +18.23% (p=0.000 n=10)
ProcessTargets/per-node/800000-32                                824.2m ±  7%   843.1m ± 5%        ~ (p=0.393 n=10)
ProcessTargets/least-weighted/800000-32                          865.0m ±  9%   880.7m ± 8%        ~ (p=0.853 n=10)
ProcessTargets/consistent-hashing/800000-32                      796.5m ±  8%   840.2m ± 5%   +5.49% (p=0.019 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/1000-32           1.517m ±  1%   1.480m ± 2%   -2.43% (p=0.002 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/1000-32       1.521m ±  2%   1.495m ± 3%   -1.73% (p=0.015 n=10)
ProcessTargetsWithRelabelConfig/per-node/1000-32                 1.522m ±  1%   1.472m ± 2%   -3.25% (p=0.002 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/10000-32          12.48m ±  4%   11.09m ± 1%  -11.15% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/10000-32      12.20m ±  1%   11.43m ± 3%   -6.32% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/10000-32                12.28m ±  3%   11.30m ± 4%   -8.02% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/100000-32         134.4m ±  2%   110.6m ± 2%  -17.71% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/100000-32     134.4m ±  4%   113.3m ± 2%  -15.72% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/100000-32               133.0m ±  2%   111.6m ± 3%  -16.07% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/800000-32        1349.0m ± 14%   936.5m ± 4%  -30.58% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/800000-32    1255.6m ±  7%   919.5m ± 3%  -26.77% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/800000-32              1407.2m ±  9%   923.0m ± 3%  -34.41% (p=0.000 n=10)
geomean                                                          32.94m         31.68m        -3.80%

                                                             │ bench_main.txt │          bench_branch.txt           │
                                                             │      B/op      │     B/op      vs base               │
ProcessTargets/least-weighted/1000-32                            4.358Mi ± 0%   4.249Mi ± 0%  -2.52% (p=0.000 n=10)
ProcessTargets/consistent-hashing/1000-32                        4.358Mi ± 0%   4.249Mi ± 0%  -2.52% (p=0.000 n=10)
ProcessTargets/per-node/1000-32                                  4.358Mi ± 0%   4.249Mi ± 0%  -2.52% (p=0.000 n=10)
ProcessTargets/least-weighted/10000-32                           43.39Mi ± 0%   42.39Mi ± 0%  -2.30% (p=0.000 n=10)
ProcessTargets/consistent-hashing/10000-32                       43.39Mi ± 0%   42.39Mi ± 0%  -2.30% (p=0.000 n=10)
ProcessTargets/per-node/10000-32                                 43.39Mi ± 0%   42.39Mi ± 0%  -2.30% (p=0.000 n=10)
ProcessTargets/least-weighted/100000-32                          433.0Mi ± 0%   424.0Mi ± 0%  -2.08% (p=0.000 n=10)
ProcessTargets/consistent-hashing/100000-32                      433.0Mi ± 0%   424.0Mi ± 0%  -2.08% (p=0.000 n=10)
ProcessTargets/per-node/100000-32                                433.0Mi ± 0%   424.0Mi ± 0%  -2.08% (p=0.000 n=10)
ProcessTargets/per-node/800000-32                                3.428Gi ± 0%   3.356Gi ± 0%  -2.10% (p=0.000 n=10)
ProcessTargets/least-weighted/800000-32                          3.428Gi ± 0%   3.356Gi ± 0%  -2.10% (p=0.000 n=10)
ProcessTargets/consistent-hashing/800000-32                      3.428Gi ± 0%   3.356Gi ± 0%  -2.10% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/1000-32           4.665Mi ± 0%   4.502Mi ± 0%  -3.48% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/1000-32       4.665Mi ± 0%   4.502Mi ± 0%  -3.48% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/1000-32                 4.664Mi ± 0%   4.502Mi ± 0%  -3.48% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/10000-32          46.44Mi ± 0%   44.97Mi ± 0%  -3.17% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/10000-32      46.44Mi ± 0%   44.97Mi ± 0%  -3.17% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/10000-32                46.44Mi ± 0%   44.97Mi ± 0%  -3.17% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/100000-32         463.5Mi ± 0%   449.8Mi ± 0%  -2.95% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/100000-32     463.5Mi ± 0%   449.9Mi ± 0%  -2.94% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/100000-32               463.5Mi ± 0%   449.8Mi ± 0%  -2.95% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/800000-32         3.667Gi ± 0%   3.535Gi ± 0%  -3.61% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/800000-32     3.667Gi ± 0%   3.535Gi ± 0%  -3.60% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/800000-32               3.667Gi ± 0%   3.535Gi ± 0%  -3.60% (p=0.000 n=10)
geomean                                                          134.7Mi        131.0Mi       -2.78%

                                                             │ bench_main.txt │          bench_branch.txt           │
                                                             │   allocs/op    │  allocs/op   vs base                │
ProcessTargets/least-weighted/1000-32                             5.416k ± 0%   5.311k ± 0%   -1.93% (p=0.000 n=10)
ProcessTargets/consistent-hashing/1000-32                         5.416k ± 0%   5.311k ± 0%   -1.94% (p=0.000 n=10)
ProcessTargets/per-node/1000-32                                   5.416k ± 0%   5.311k ± 0%   -1.94% (p=0.000 n=10)
ProcessTargets/least-weighted/10000-32                            53.99k ± 0%   52.97k ± 0%   -1.88% (p=0.000 n=10)
ProcessTargets/consistent-hashing/10000-32                        53.99k ± 0%   52.97k ± 0%   -1.88% (p=0.000 n=10)
ProcessTargets/per-node/10000-32                                  53.99k ± 0%   52.97k ± 0%   -1.88% (p=0.000 n=10)
ProcessTargets/least-weighted/100000-32                           539.7k ± 0%   528.5k ± 0%   -2.07% (p=0.000 n=10)
ProcessTargets/consistent-hashing/100000-32                       539.7k ± 0%   528.5k ± 0%   -2.07% (p=0.000 n=10)
ProcessTargets/per-node/100000-32                                 539.7k ± 0%   528.5k ± 0%   -2.07% (p=0.000 n=10)
ProcessTargets/per-node/800000-32                                 4.324M ± 0%   4.230M ± 0%   -2.16% (p=0.000 n=10)
ProcessTargets/least-weighted/800000-32                           4.325M ± 0%   4.230M ± 0%   -2.18% (p=0.001 n=10)
ProcessTargets/consistent-hashing/800000-32                       4.324M ± 0%   4.230M ± 0%   -2.17% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/1000-32            8.419k ± 0%   7.303k ± 0%  -13.26% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/1000-32        8.420k ± 0%   7.303k ± 0%  -13.27% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/1000-32                  8.419k ± 0%   7.303k ± 0%  -13.26% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/10000-32           83.99k ± 0%   72.93k ± 0%  -13.17% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/10000-32       83.99k ± 0%   72.93k ± 0%  -13.17% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/10000-32                 83.99k ± 0%   72.93k ± 0%  -13.17% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/100000-32          839.7k ± 0%   728.3k ± 0%  -13.26% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/100000-32      839.7k ± 0%   728.3k ± 0%  -13.26% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/100000-32                839.6k ± 0%   728.3k ± 0%  -13.26% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/800000-32          6.726M ± 0%   5.827M ± 0%  -13.36% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/800000-32      6.725M ± 0%   5.827M ± 0%  -13.35% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/800000-32                6.726M ± 0%   5.827M ± 0%  -13.36% (p=0.000 n=10)
geomean                                                           201.6k        185.8k        -7.81%
```